move game settings to settings mod
