+++ src/main.rs
@@ -5,6 +5,7 @@ use dimension_macro_derive::Enumerated;
 
 mod menu;
 mod game;
+mod settings;
 
 #[derive(States, Component, Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
 enum AppState {
@@ -12,65 +13,6 @@ enum AppState {
     Menu,
     Game,
 }
-
-// A Setting is any enum which (1) has a variants() method, (2) can be Displayed, and (3) is a Component
-trait Setting: std::fmt::Display + Component + Clone + Copy {}
-
-#[derive(Resource, Component, Enumerated, Clone, Copy, Default, PartialEq, Eq)]
-enum HumanMark {
-    #[default]
-    HumanX,
-    HumanO
-}
-
-impl std::fmt::Display for HumanMark {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", match self {
-            HumanMark::HumanX => "Human X",
-            HumanMark::HumanO => "Human O",
-        })
-    }
-}
-
-impl Setting for HumanMark {}
-
-#[derive(Resource, Component, Enumerated, Clone, Copy, Default, PartialEq, Eq)]
-enum Difficulty {
-    Easy,
-    Medium,
-    #[default]
-    Hard,
-}
-
-impl std::fmt::Display for Difficulty {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", match self {
-            Difficulty::Easy => "Easy",
-            Difficulty::Medium => "Medium",
-            Difficulty::Hard => "Hard",
-        })
-    }
-}
-
-impl Setting for Difficulty {}
-
-#[derive(Resource, Component, Enumerated, Clone, Copy, Default, PartialEq, Eq)]
-enum GameMode {
-    OnePlayer,
-    #[default]
-    TwoPlayers,
-}
-
-impl std::fmt::Display for GameMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", match self {
-            GameMode::OnePlayer => "One Player",
-            GameMode::TwoPlayers => "Two Players",
-        })
-    }
-}
-
-impl Setting for GameMode {}
 
 trait Enumerated {
     type Item;
@@ -81,9 +23,9 @@ trait Enumerated {
 fn main() {
     App::new()
         .insert_resource(AssetMetaCheck::Never) // https://github.com/bevyengine/bevy/issues/10157#issuecomment-1849092112
-        .insert_resource(GameMode::default())
-        .insert_resource(HumanMark::default())
-        .insert_resource(Difficulty::default())
+        .insert_resource(settings::GameMode::default())
+        .insert_resource(settings::HumanMark::default())
+        .insert_resource(settings::Difficulty::default())
         .add_plugins(DefaultPlugins)
         .insert_resource(ClearColor(Color::rgb(0.9, 0.9, 0.9)))
         .init_state::<AppState>()
